fix for brightness

+++ Common/POCOs/Shots/SampleData.cs
@@ -8,7 +8,7 @@ public class SampleData: Poco
 {
     public SampleData() { }
 
-    public SampleData(string? type, float? timestamp, int count, float brightness, double? x, double? y, double? z)
+    public SampleData(string? type, float? timestamp, int count, double brightness, double? x, double? y, double? z)
     {
         Type = type;
         Logtime = timestamp;
@@ -27,7 +27,7 @@ public class SampleData: Poco
     [JsonPropertyName("Count")]
     public int? Count { get; set; }
     [JsonPropertyName("Brightness")]
-    public float? Brightness { get; set; }
+    public double? Brightness { get; set; }
 
 
     [JsonPropertyName("X")]
